✨ MagazineViewModel 과 MagazineDetailViewModel 분리 ( 북마크 로직 수정 )
